move JUnitTest* structs to public package
